Updated docker-entrypoint.sh to avoid installation loop.

+++ docker-entrypoint.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
-# If MyBB zip file is found, process.
-if [ ! -f /tmp/mybb.zip ];
+# If MyBB zip file is not found & no index, begin installation process.
+if [ ! -f /tmp/mybb.zip ] && [ ! -f /var/www/html/index.php];
 then
     curl -o /tmp/mybb.zip -fSL https://resources.mybb.com/downloads/mybb_1820.zip;
     unzip /tmp/mybb.zip -d /var/www/html/;
@@ -10,9 +10,6 @@ then
     rm -rfv /var/www/html/Upload;
     rm -rfv /var/www/html/Documentation;
     rm -rfv /tmp/mybb.zip;
-elif [ -f /var/www/html/index.php ];
-then
-    echo "MyBB files present. Assuming its installed already...";
 fi
 
 # Start Apache
